moving auth and api config
