+++ packages/amplify/amplify_flutter/pubspec.yaml
@@ -18,7 +18,6 @@ dependencies:
   collection: ^1.15.0
   flutter:
     sdk: flutter
-  json_annotation: ^4.3.0
   meta: ^1.7.0
   plugin_platform_interface: ^2.0.0
