+++ packages/amplify_core/lib/src/types/config/amplify_plugin_config.dart
@@ -13,7 +13,9 @@
 // permissions and limitations under the License.
 //
 
-import 'package:amplify_flutter/amplify_flutter.dart';
+import 'package:amplify_core/src/types/config/amplify_plugin_registry.dart';
+import 'package:amplify_core/src/util/equatable.dart';
+import 'package:amplify_core/src/util/serializable.dart';
 import 'package:collection/collection.dart';
 import 'package:meta/meta.dart';
