+++ packages/amplify_core/lib/src/types/config/amplify_plugin_registry.dart
