+++ packages/amplify_core/lib/src/types/config/api/appsync/api_config.dart
@@ -13,7 +13,9 @@
 // permissions and limitations under the License.
 //
 
-import 'package:amplify_flutter/amplify_flutter.dart';
+import 'package:amplify_core/src/types/config/api/api_config.dart';
+import 'package:amplify_core/src/util/equatable.dart';
+import 'package:amplify_core/src/util/serializable.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 part 'api_config.g.dart';
