+++ packages/amplify_core/lib/src/types/config/api/appsync/endpoint_type.dart
