+++ packages/amplify_core/lib/src/types/config/api/aws_api_config.dart
@@ -13,12 +13,13 @@
 // permissions and limitations under the License.
 //
 
-import 'package:amplify_flutter/amplify_flutter.dart';
+import 'package:amplify_core/src/types/config/amplify_plugin_config.dart';
+import 'package:amplify_core/src/types/config/amplify_plugin_registry.dart';
+import 'package:amplify_core/src/types/config/api/api_config.dart';
+import 'package:amplify_core/src/types/config/config_map.dart';
 import 'package:collection/collection.dart';
 import 'package:meta/meta.dart';
 
-export 'package:amplify_api_plugin_interface/amplify_api_plugin_interface.dart'
-    show APIAuthorizationType;
 export 'appsync/api_config.dart';
 export 'appsync/endpoint_type.dart';
