+++ packages/amplify_core/lib/src/types/config/auth/auth_config.dart
@@ -13,9 +13,13 @@
 // permissions and limitations under the License.
 //
 
-import 'package:amplify_flutter/amplify_flutter.dart';
-
 export 'cognito_config.dart' hide CognitoPluginConfigFactory;
+
+import 'package:amplify_core/src/types/config/amplify_plugin_config.dart';
+import 'package:amplify_core/src/types/config/amplify_plugin_registry.dart';
+import 'package:amplify_core/src/types/config/auth/auth_config.dart';
+import 'package:amplify_core/src/types/config/config_map.dart';
+import 'package:amplify_core/src/util/serializable.dart';
 
 part 'auth_config.g.dart';
