+++ packages/amplify_core/lib/src/types/config/auth/cognito/auth.dart
