+++ packages/amplify_core/lib/src/types/config/auth/cognito/authentication_flow_type.dart
