+++ packages/amplify_core/lib/src/types/config/auth/cognito/password_protection_settings.dart
@@ -13,7 +13,8 @@
 // permissions and limitations under the License.
 //
 
-import 'package:amplify_flutter/amplify_flutter.dart';
+import 'package:amplify_core/src/util/equatable.dart';
+import 'package:amplify_core/src/util/serializable.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 part 'password_protection_settings.g.dart';
