+++ packages/amplify_core/lib/src/types/config/auth/cognito/pinpoint_analytics.dart
