+++ packages/amplify_core/lib/src/types/config/auth/cognito/s3_transfer_utility.dart
