+++ packages/amplify_core/lib/src/types/config/auth/cognito/social_provider.dart
