+++ packages/amplify_core/lib/src/types/config/auth/cognito/user_attribute_key.dart
