+++ packages/amplify_core/lib/src/types/config/config_map.dart
@@ -15,7 +15,10 @@
 
 import 'dart:collection';
 
-import 'package:amplify_flutter/amplify_flutter.dart';
+import 'package:amplify_core/src/types/config/amplify_plugin_config.dart';
+import 'package:amplify_core/src/types/config/amplify_plugin_registry.dart';
+import 'package:amplify_core/src/util/equatable.dart';
+import 'package:amplify_core/src/util/serializable.dart';
 import 'package:collection/collection.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:meta/meta.dart';
