+++ packages/amplify_core/lib/src/util/equatable.dart
