+++ packages/amplify_core/pubspec.yaml
@@ -13,6 +13,7 @@ dependencies:
   plugin_platform_interface: ^2.0.0
   collection: ^1.15.0
   date_time_format: ^2.0.1
+  json_annotation: ^4.3.0
   meta: ^1.3.0
   uuid: ^3.0.1
